0.2.0
-Full functionality for straight-line projectile weapons:
Weapon data is stored in a scriptable object. The WeaponHandler script takes weapon data and executes the Fire() method on the scriptable object (ProjectileWeapon). This is an override method which will carry out functions like instantiating a prefab projectile or, on future hitscan weapons, firing a raycast.

For the projectile weapon, it calls a method on InstantiateProjectile, passing in the projectile game object stored on the scriptable object. All the weapon does is say which prefab to spawn and how often (clip size, fire rate e.c.t.). Unique behaviours can be added to the projectile itself.

-Each weapon's remaining ammo and reload time is stored independently

-Added basic UI and player UI handler

-Added health script to be used on both the player and enemies